requirements.txt
polars>=0.20.0
pandas>=2.0.0
click>=8.0.0
pydantic>=2.0.0
rich>=13.0.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
